move css rule to a separate folder
