+++ packages/css/src/rules/S5362/index.ts
@@ -17,17 +17,4 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-import * as stylelint from 'stylelint';
-import { rule as S5362 } from './S5362';
-
-/**
- * The set of internal Stylelint-based rules
- */
-const rules: { [key: string]: stylelint.Rule<any, any> } = {};
-
-/**
- * Maps Stylelint rule keys to rule implementations
- */
-rules[S5362.ruleName] = S5362.rule; // function-calc-no-invalid
-
-export { rules };
+export { rule } from './rule';
